pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PyAderlee"
version = "1.0"
description = "A lightweight Python library for data processing and file handling"
readme = "README.md"
authors = [
    { name = "Khaled Karman", email = "k@rawasy.com" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["database", "sqlite", "file handling", "data processing"]
dependencies = [
    "pathlib>=1.0",
    "requests>=2.25.0",
    "mysql-connector-python>=8.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
dev = [
    "black>=22.0",
    "isort>=5.0",
    "mypy>=1.0",
    "flake8>=5.0",
]

[project.urls]
Homepage = "https://github.com/rawasy/PyAderlee"
Documentation = "https://github.com/rawasy/PyAderlee#readme"
Issues = "https://github.com/rawasy/PyAderlee/issues"

[tool.hatch.build.targets.wheel]
packages = ["PyAderlee"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=PyAderlee"
testpaths = [
    "tests",
] 